requirements.txt
Pillow~=9.5.0
opencv-python>=4.0.0.0
requests>=2.25.1
emoji==2.2.0
jishaku
git+https://github.com/Rapptz/asqlite
git+https://github.com/Rapptz/discord.py
git+https://github.com/Rapptz/discord-ext-menus
git+https://github.com/oliver-ni/discord-ext-menus-views
charset-normalizer
colour-science
numpy<2
aiohttp
setuptools
visual-center
more_itertools
tldextract~=3.4.0

attrs~=22.2.0
cpmpy

tomlkit
